refactor: Move application logic to app module and simplify main function
